+++ contracts/market/src/constants.rs
@@ -32,19 +32,19 @@ pub const SECONDS_PER_LEDGER: u64 = 6;
 pub const LEDGERS_PER_DAY: u32 = (SECONDS_PER_DAY / SECONDS_PER_LEDGER) as u32;
 
 /// Instance storage extension is spread among all users, so it must be cheap and paid regularly
-pub const INSTANCE_THRESHOLD: u32 = 40 * (LEDGERS_PER_DAY as u32);
-pub const INSTANCE_BUMP: u32 = INSTANCE_THRESHOLD + (LEDGERS_PER_DAY as u32);
+pub const INSTANCE_THRESHOLD: u32 = 40 * LEDGERS_PER_DAY;
+pub const INSTANCE_BUMP: u32 = INSTANCE_THRESHOLD + LEDGERS_PER_DAY;
 
 /// Shared persistent storage extension is spread among all shared resource users, so it must be
 /// cheap and paid regularly
-pub const SHARED_THRESHOLD: u32 = 50 * (LEDGERS_PER_DAY as u32);
-pub const SHARED_BUMP: u32 = SHARED_THRESHOLD + (LEDGERS_PER_DAY as u32);
+pub const SHARED_THRESHOLD: u32 = 50 * LEDGERS_PER_DAY;
+pub const SHARED_BUMP: u32 = SHARED_THRESHOLD + LEDGERS_PER_DAY;
 
 /// Individual persistent storage extension is usually paid by the data owners. It should neither be
 /// paid very often (to reduce extension operation costs) nor very rarely (to minimize archival
 /// risk)(TODO: Though, is it really a sound argument?)
-pub const INDIVIDUAL_THRESHOLD: u32 = 160 * (LEDGERS_PER_DAY as u32);
-pub const INDIVIDUAL_BUMP: u32 = 180 * (LEDGERS_PER_DAY as u32);
+pub const INDIVIDUAL_THRESHOLD: u32 = 160 * LEDGERS_PER_DAY;
+pub const INDIVIDUAL_BUMP: u32 = 180 * LEDGERS_PER_DAY;
 
 // ---- Interest Rate and Accrual ----
 
@@ -124,7 +124,7 @@ pub const MIN_LEVERAGE_MULTIPLIER: u32 = LEVERAGE_SCALE;
 
 /// Maximum acceptable oracle price age in seconds
 /// TODO: How to properly pick this value?
-pub const MAX_ORACLE_PRICE_AGE_SECONDS: u64 = (15 * SECONDS_PER_MINUTE) as u64;
+pub const MAX_ORACLE_PRICE_AGE_SECONDS: u64 = 15 * SECONDS_PER_MINUTE;
 
 // ---- Contract Addresses ----
